Move mantra editing to its own page
It was kind of obnoxious having everything in text inputs, this moves editing and deleting to its own page.

Also moved the form to create a new mantra to the top of the page so you don't have to scroll past your other mantras every time you want to make a new one.